move TelemetryReader to stabilizer module
