+++ py/setup.py
@@ -8,5 +8,6 @@ setup(name="stabilizer",
       license="MIT",
       install_requires=[
             "numpy",
+            "gmqtt",
             "miniconf-mqtt@git+https://github.com/quartiq/miniconf@develop#subdirectory=py/miniconf-mqtt"
       ])
